Fixing authentication errors in GitHub API methods.
Retrieving and cloning  installation repos now requires the `app` auth type.
Same change had to be made for authenticating with an app token.

The JWT expiry was lowered to 5 minutes which is fine because these are very short-lived.
The max expiry is 10 minutes which seems not to work with the GitHub API.